# docs/swagger.yaml
openapi: 3.0.3
info:
  title: AIDOC API
  description: |
    AI Health Tracker API - Multi-protocol architecture with resilience patterns:
    - **GraphQL** (`/api/graphql`) - Web app queries
    - **SSE** (`/api/chat`) - AI chat streaming
    - **REST** (`/api/health`, `/api/oura`) - Health checks & Oura integration
    - **gRPC** (`:50051`) - Mobile/Watch clients

    **Resilience Features:**
    - Circuit breakers for external services
    - Automatic fallbacks (LiteLLM)
    - Response caching
    - Retry with exponential backoff
  version: 0.0.1
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Local HTTP
  - url: grpc://localhost:50051
    description: Local gRPC
paths:
...
  /oura:
    get:
      summary: Fetch Oura Ring health data
      operationId: getOuraData
      tags: [Oura Ring]
      description: Returns sleep, activity, readiness data from Oura Cloud API
      responses:
        "200":
          description: Health data from Oura Ring
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OuraDataResponse"
        "401":
          description: Not authenticated with Oura
        "500":
          description: Oura API error
...
  /oura/callback:
    get:
      summary: OAuth2 callback from Oura
      operationId: ouraCallback
      tags: [Oura Ring]
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect to app after successful auth
        "400":
          description: Invalid state or code
components:
  schemas:
...
    OuraDataResponse:
      type: object
      properties:
        sleep:
          type: array
          items:
            type: object
            properties:
              day: { type: string }
              score: { type: integer }
              totalSleep: { type: number }
        activity:
          type: array
          items:
            type: object
            properties:
              day: { type: string }
              steps: { type: integer }
              calories: { type: integer }
        readiness:
          type: array
          items:
            type: object
            properties:
              day: { type: string }
              score: { type: integer }
...